EnderecoController autorização
Somente secretário pode acessar/modificar

+++ controllers/EnderecoController.php
@@ -35,13 +35,24 @@ class EnderecoController extends Controller
      */
     public function actionIndex()
     {
-        $searchModel = new EnderecoSearch();
-        $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
+        if(!Yii::$app->user->isGuest){
+			if(Yii::$app->user->identity->id_Yii == 2)
+			{
+                $searchModel = new EnderecoSearch();
+                $dataProvider = $searchModel->search(Yii::$app->request->queryParams);
 
-        return $this->render('index', [
-            'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider,
-        ]);
+                return $this->render('index', [
+                    'searchModel' => $searchModel,
+                    'dataProvider' => $dataProvider,
+                ]);
+            }
+            else{
+                throw new NotFoundHttpException(Yii::t('app', 'Page not found.'));
+            }
+        }
+        else{
+			return $this->redirect(['site/login']);
+		}
     }
 
     /**
@@ -52,9 +63,20 @@ class EnderecoController extends Controller
      */
     public function actionView($id)
     {
-        return $this->render('view', [
-            'model' => $this->findModel($id),
-        ]);
+        if(!Yii::$app->user->isGuest){
+			if(Yii::$app->user->identity->id_Yii == 2)
+			{
+                return $this->render('view', [
+                    'model' => $this->findModel($id),
+                ]);
+            }
+            else{
+                throw new NotFoundHttpException(Yii::t('app', 'Page not found.'));
+            }
+        }
+        else{
+			return $this->redirect(['site/login']);
+		}
     }
 
     /**
@@ -64,15 +86,26 @@ class EnderecoController extends Controller
      */
     public function actionCreate()
     {
-        $model = new Endereco();
+        if(!Yii::$app->user->isGuest){
+			if(Yii::$app->user->identity->id_Yii == 2)
+			{
+                $model = new Endereco();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id_usuario]);
+                if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                    return $this->redirect(['view', 'id' => $model->id_usuario]);
+                }
+
+                return $this->render('create', [
+                    'model' => $model,
+                ]);
+            }
+            else{
+                throw new NotFoundHttpException(Yii::t('app', 'Page not found.'));
+            }
         }
-
-        return $this->render('create', [
-            'model' => $model,
-        ]);
+        else{
+			return $this->redirect(['site/login']);
+		}
     }
 
     /**
@@ -84,15 +117,26 @@ class EnderecoController extends Controller
      */
     public function actionUpdate($id)
     {
-        $model = $this->findModel($id);
+        if(!Yii::$app->user->isGuest){
+			if(Yii::$app->user->identity->id_Yii == 2)
+			{
+                $model = $this->findModel($id);
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id_usuario]);
+                if ($model->load(Yii::$app->request->post()) && $model->save()) {
+                    return $this->redirect(['view', 'id' => $model->id_usuario]);
+                }
+
+                return $this->render('update', [
+                    'model' => $model,
+                ]);
+            }
+            else{
+                throw new NotFoundHttpException(Yii::t('app', 'Page not found.'));
+            }
         }
-
-        return $this->render('update', [
-            'model' => $model,
-        ]);
+        else{
+			return $this->redirect(['site/login']);
+		}
     }
 
     /**
@@ -104,9 +148,20 @@ class EnderecoController extends Controller
      */
     public function actionDelete($id)
     {
-        $this->findModel($id)->delete();
+        if(!Yii::$app->user->isGuest){
+			if(Yii::$app->user->identity->id_Yii == 2)
+			{
+                $this->findModel($id)->delete();
 
-        return $this->redirect(['index']);
+                return $this->redirect(['index']);
+            }
+            else{
+                throw new NotFoundHttpException(Yii::t('app', 'Page not found.'));
+            }
+        }
+        else{
+			return $this->redirect(['site/login']);
+		}
     }
 
     /**
